cleanup, simplify diffing
