--- src/store/redux-bridge/strategies/deepDiff/arrayDiff/index.js
@@ -1,10 +1,0 @@
-import same from "./diff/same";
-
-export * from "./diff/diff";
-export {
-  same
-};
-
-export * from "./diff/patch";
-
-export * from "./diff/apply";
