+++ src/store/redux-bridge/utils/constants.js
@@ -3,7 +3,3 @@ export const DIFF_STATUS_UPDATED = 'updated';
 
 // The `change` value for removed fields resulting from shallow diff
 export const DIFF_STATUS_REMOVED = 'removed';
-
-export const DIFF_STATUS_KEYS_UPDATED = 'updated_keys';
-
-export const DIFF_STATUS_ARRAY_UPDATED = 'updated_array';
